Adding "Adaptive" label to all printed outputs of AMPI. Fixing integer overflow problems in multiple kernels. Cutting AMPI execution of AMPI DGEMM short, pending fix of its MPI_Comm_create with multiple groups. Adjusting size of problems to be run in the "wide" category.

+++ FG_MPI/Reduce/reduce.c
@@ -73,7 +73,7 @@ int main(int argc, char ** argv)
   int my_ID;            /* rank                                              */
   int root=0;
   int iterations;       /* number of times the reduction is carried out      */
-  int i, iter;          /* dummies                                           */
+  long i, iter;         /* dummies                                           */
   long vector_length;   /* length of the vectors to be aggregated            */
   double * RESTRICT vector; /* vector to be reduced                          */
   double * RESTRICT ones;   /* constant vector                               */
